feat: 基本功能全部实现 todo: 支持reload命令、其他的认证方式等

+++ EtherosAuthPaper/src/main/java/cn/skilfully/etheros/config/entity/PluginConfig.java
@@ -63,9 +63,9 @@ public class PluginConfig {
                 @Setting
                 private Double z;
                 @Setting
-                private Double yaw;
+                private Float yaw;
                 @Setting
-                private Double pitch;
+                private Float pitch;
                 @Setting("auto-return")
                 private Boolean autoReturn;
             }
